+++ test/compile/mark/encode/position-point.test.ts
@@ -1,4 +1,4 @@
-import {getSizeChannel, X, Y} from '../../../../src/channel';
+import {X, Y} from '../../../../src/channel';
 import {pointPosition} from '../../../../src/compile/mark/encode';
 import {parseUnitModelWithScaleAndLayoutSize} from '../../../util';
 
@@ -27,40 +27,6 @@ describe('compile/mark/encode/position-point', () => {
     [X, Y].forEach(channel => {
       const mixins = pointPosition(channel, model, {defaultPos: 'zeroOrMin'});
       expect(mixins[channel]['field']).toEqual(model.getName(channel));
-    });
-  });
-
-  it('should return correctly for lat/lng with size', () => {
-    const model = parseUnitModelWithScaleAndLayoutSize({
-      data: {
-        url: 'data/zipcodes.csv',
-        format: {
-          type: 'csv'
-        }
-      },
-      mark: {
-        type: 'image',
-        width: 42,
-        height: 42
-      },
-      encoding: {
-        longitude: {
-          field: 'longitude',
-          type: 'quantitative'
-        },
-        latitude: {
-          field: 'latitude',
-          type: 'quantitative'
-        }
-      }
-    });
-
-    [X, Y].forEach(channel => {
-      const mixins = pointPosition(channel, model, {defaultPos: 'zeroOrMin'});
-      expect(mixins[channel]['field']).toEqual(model.getName(channel));
-
-      const sizeChannel = getSizeChannel(channel);
-      expect(mixins[sizeChannel]).toEqual({[sizeChannel]: 42});
     });
   });
 });
